+++ autolens/plot/fit_point_plotters.py
@@ -2,7 +2,7 @@ from autoarray.plot.abstract_plotters import AbstractPlotter
 
 import autogalaxy.plot as aplt
 
-from autolens.fit.fit_point import FitPointDataset
+from autolens.point.fit_point import FitPointDataset
 
 
 class FitPointDatasetPlotter(AbstractPlotter):
