--- test_autolens/positions_solver/centred_00/pickles/.gitignore
@@ -1,2 +1,0 @@
-*
-!.gitignore
